src/backend/requirements.txt
django>=3.2,<5.0
djangorestframework
django-cors-headers
psycopg2-binary
Pillow
python-dotenv
djangorestframework_simplejwt
faker